requirements.txt
fastapi
granian[reload]
requests
pydantic
pydantic-settings
pydantic-core
typing-inspection
fake-useragent
brotli
httpx
python-dotenv
